feat: time to production chart
